[Feature] IBInspectable
Removed views for single styles in favor of one class with inspectable attributes to set those in IB directly.

The benefit would be to have fewer classes and give designers the ability to modify styles on their own without having to read any of the code.

Also added some validation of text fields, changing attributes of labels and text fields when invalid input is detected and disabling button (inverting background color/font color to indicate disabled state - there might be better ways to do that).